Implemented Full Waypoint Walkthrough
Implemented the code outlined in the Udacity's self driving's car Full Waypoint Walkthrough lesson.

Simulated car is able to follow waypoints and decelerate to and stop at red light.
Required to enable the "Camera" checkbox and disable the "Manual" checkbox in the vehicle simulator to work, as the full waypoint update depends on subscribing the traffic light state determined by `tl_detector.py`